Initial: Display temperature and humidity
This new app has the ability to display hourly weather metrics for a
given location in the world. It uses the open-meteo API. The app will
support choosing any two parameters provided by the open-meteo API, but
currently only temperature and humidity have been implemented.

Still working through readability of axis labels / the whole thing. It's
a bit of a mess, but I still find it useful!